requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
streamlit>=1.35.0
requests>=2.31.0
python-dotenv>=1.0.0
discord.py>=2.3.0
slack-sdk>=3.27.0
aiohttp>=3.9.0
